fix: resolve circular dependencies in council modules
- Move getSearchAddress function from lib/addressExtractor to councils/core/addressFormatter
- Update utils.ts to import directly from specific modules instead of index
- Update all council implementations to use new addressFormatter import
- Eliminate circular dependency warnings between addressExtractor and council modules
