+++ convex/councils/core/index.ts
@@ -1,5 +1,6 @@
 // Re-export all core utilities from a single location
 
+export * from "./addressFormatter";
 export * from "./errors";
 export * from "./types";
 export * from "./utils";
